projects/ai-newsletter/requirements.txt
requests
python-dotenv
pytest>=7.0.0
pytest-mock>=3.10.0
